+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Added
+- added support for running test in project where wtr.plugin.runner.js is present, plugin will call this file with given arguments to run tests in single file and by pattern like via peonRunner.js
+
 ## 2.3.0
 
 ### Fixed